move UserService to common module as it is needed by gateway app
 - also rename AddTokenHeaderFilter to AddAccessTokenCookieFilter
