+++ gateway/src/main/java/com/quizme/GatewayApplication.java
@@ -1,6 +1,6 @@
 package com.quizme;
 
-import com.quizme.auth.AddTokenHeaderFilter;
+import com.quizme.auth.AddAccessTokenCookieFilter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -30,11 +30,11 @@ public class GatewayApplication {
 
     @Bean
     public RouterFunction<ServerResponse> myRoutes(@Value("${app.backend_url}") String backendUrl,
-                                                   AddTokenHeaderFilter addTokenHeaderFilter) {
+                                                   AddAccessTokenCookieFilter addAccessTokenCookieFilter) {
         return route("my_backend")
                 .route(path("/**"), http())
                 .before(uri(backendUrl))
-                .filter(addTokenHeaderFilter)
+                .filter(addAccessTokenCookieFilter)
                 .build();
     }
